backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
mistralai==0.0.10
python-magic==0.4.27
python-docx==0.8.11
PyPDF2==3.0.1 